+++ packages/orq-rc/src/models/components/updatemcpgatewayrequest.ts
@@ -9,6 +9,11 @@ import {
   McpGatewayMode$outboundSchema,
 } from "./mcpgatewaymode.js";
 import {
+  McpGatewayPlugin,
+  McpGatewayPlugin$Outbound,
+  McpGatewayPlugin$outboundSchema,
+} from "./mcpgatewayplugin.js";
+import {
   McpGatewayServerLink,
   McpGatewayServerLink$Outbound,
   McpGatewayServerLink$outboundSchema,
@@ -59,6 +64,10 @@ export type UpdateMcpGatewayRequest = {
    * Set true to remove every link; cannot be combined with `server_links`.
    */
   clearServerLinks?: boolean | undefined;
+  /**
+   * Plugins run on every tool call this gateway serves.
+   */
+  plugins?: Array<McpGatewayPlugin> | undefined;
 };
 
 /** @internal */
@@ -72,6 +81,7 @@ export type UpdateMcpGatewayRequest$Outbound = {
   mode?: string | undefined;
   sharing?: Sharing$Outbound | undefined;
   clear_server_links?: boolean | undefined;
+  plugins?: Array<McpGatewayPlugin$Outbound> | undefined;
 };
 
 /** @internal */
@@ -89,6 +99,7 @@ export const UpdateMcpGatewayRequest$outboundSchema: z.ZodType<
   mode: McpGatewayMode$outboundSchema.optional(),
   sharing: Sharing$outboundSchema.optional(),
   clearServerLinks: z.boolean().optional(),
+  plugins: z.array(McpGatewayPlugin$outboundSchema).optional(),
 }).transform((v) => {
   return remap$(v, {
     displayName: "display_name",