+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3898,7 +3898,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M1KMERW4WWM5G7WYFQ98H5DV"),
+  id: z.string().default("01M1NQ4PMH23SM3S5WSKMEFRA0"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),