+++ packages/orq-rc/RELEASES.md
@@ -12089,3 +12089,13 @@ Based on:
 - [typescript v4.15.0-rc.19] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.19 - packages/orq-rc
+
+## 2026-09-04 08:09:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.21] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.21 - packages/orq-rc